requirements.txt
# --- Core Web Framework ---
django==5.2.11
dj-database-url==3.1.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.11.1
sqlparse==0.5.5
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1
pyjwt==2.11.0

# --- Background Tasks & Caching ---
celery==5.6.2
redis==7.2.0
django-celery-results==2.6.0

# --- Websocket Packages ---
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1

# --- AI & LLM (Models & Frameworks) ---
langchain==1.2.10
langchain-core==1.2.14
langchain-community==0.4.1
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-huggingface==1.2.0
groq==0.37.1
huggingface-hub==0.36.2

# --- Multi-Agent Orchestration ---
langgraph==1.0.9

# --- Vector Database & PDF Processing ---
qdrant-client==1.17.0

# --- Environment & Utilities ---
python-dotenv==1.2.1
requests==2.32.5

# Brevo SMTP credentials (for email functionality)
django-anymail==14.0

# Google OAuth dependencies
google-auth==2.20.0


# Extraction related dependencies
pypdf2==3.0.1

# Adding shared model
git+https://github.com/dhruv9512/docugyan-shared-models.git@main#egg=docugyan-shared-models